+++ app/src/main/java/com/byeboo/app/presentation/quest/record/QuestRecordingState.kt
@@ -8,8 +8,8 @@ import com.byeboo.app.domain.model.QuestWritingState
 data class QuestRecordingState(
     val questId: Long = 0,
     val step: String = "감정 쏟아내기",
-    val stepNumber: Int = 0,
-    val questNumber: Int = 0,
+    val stepNumber: Long = 0,
+    val questNumber: Long = 0,
     val questQuestion: String = "연애에서 반복됐던 문제 패턴 3가지를 생각해보아요.",
     val questAnswer: String = "",
     val contentsState: QuestWritingState = QuestWritingState.BeforeWriting,
